redit file structure again
