+++ .gitignore
@@ -1,2 +1,4 @@
 .env
 .flaskenv
+__pycache__
+flask_skele.py
