+++ src/models/auth.py
@@ -1,4 +1,4 @@
-from . import db
+from src.main import db
 from sqlalchemy.dialects.postgresql import JSON
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
