+++ src/routes/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+try:
+    from .auth import *
+except ImportError:
+    from auth import *
